requirements.txt
os
sqlite3
smtplib
flask
email.mime.text
werkzeug.security
random